Teste input search
