--- src/components/AppContainer.js
@@ -1,24 +1,0 @@
-import React from "react"
-import styled from 'styled-components'
-import AppBar from './AppBar/AppBar'
-import SideBar from './SideBar/SideBar'
-import BodyProducts from './BodyProducts'
-
-const MainContent = styled.div`
-  display: grid;
-  grid-template-columns: 270px 1fr;
-`
-export default class AppContainer extends React.Component {
-
-  render () {
-    return (
-     <div>
-       <AppBar/>
-       <MainContent>
-          <SideBar/>
-          <BodyProducts/>
-       </MainContent>
-     </div>
-    )
-  }
-}
